rename package appmiddleware to middleware
