+++ services/api/middleware/auth.go
@@ -1,4 +1,4 @@
-package appmiddleware
+package middleware
 
 import (
 	"github.com/labstack/echo/v4"
