+++ services/api/middleware/ctxExtention.go
@@ -1,4 +1,4 @@
-package appmiddleware
+package middleware
 
 import (
 	"context"
